pyproject.toml
[tool.poetry]
name = "stats-arb"
version = "0.1.0"
description = "Statistical arbitrage trading strategy using convex concave programming."
authors = ["Kasper Johansson, Thomas Schmelzer, and Stephen Boyd"]
readme = "README.md"
packages = [{include = "cvx"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.2"
cvxpy-base = "*"
#pandas = "^2.0.0"
matplotlib = "^3.7.1"
clarabel = "^0.4.1"
cvxsimulator = "0.2.3"

[tool.poetry.dev-dependencies]
pytest = "7.1.3"
pytest-cov = "*"
jupyter = "^1.0.0"
seaborn = "^0.12.2"
dask = "^2023.3.2"
tqdm = "^4.65.0"
#
#mosek = "^10.0.40"
yfinance = "^0.2.14"
wrds = "^3.1.6"
twelvedata = "^1.2.11"
graphviz = "^0.20.1"
quantstats = "*"
loguru = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
